# optimize memory allocation