--- a/sheet_test.go
+++ b/sheet_test.go
@@ -346,75 +346,20 @@ func TestSetSheetName(t *testing.T) {
 	assert.Equal(t, "Sheet1", f.GetSheetName(0))
 }
 
-func TestNewSheetWithRowNum(t *testing.T) {
-	f := NewFile()
-	f.NewSheetWithRowNum("sheet1", 100)
-	f.SetCellInt("sheet1", "A100", 100)
-	val, err := f.GetCellValue("sheet1", "A100")
-	if err != nil {
-		t.Fatal(err)
-	}
-	assert.Equal(t, "100", val)
-}
-
-func BenchmarkNewSheetWithRowNum(b *testing.B) {
+func BenchmarkNewSheet(b *testing.B) {
 	b.RunParallel(func(pb *testing.PB) {
 		for pb.Next() {
-			newSheetWithRowNum()
-		}
-
-	})
-}
-func BenchmarkNewSheetWithOutRowNum(b *testing.B) {
-	b.RunParallel(func(pb *testing.PB) {
-		for pb.Next() {
-			newSheetWithOutRowNum()
+			newSheetWithSet()
 		}
 	})
 }
-func BenchmarkNewSheetWithStreamWriter(b *testing.B) {
-	b.RunParallel(func(pb *testing.PB) {
-		for pb.Next() {
-			newSheetWithStreamWriter()
-		}
-	})
-}
-func newSheetWithRowNum() {
-	file := NewFile()
-
-	file.NewSheetWithRowNum("sheet1", 1000)
-
-	for i := 0; i < 1000; i++ {
-		file.SetCellInt("sheet1", "A"+strconv.Itoa(i+1), i)
-	}
-	file = nil
-}
-func newSheetWithOutRowNum() {
+func newSheetWithSet() {
 	file := NewFile()
 	file.NewSheet("sheet1")
 	for i := 0; i < 1000; i++ {
 		file.SetCellInt("sheet1", "A"+strconv.Itoa(i+1), i)
 	}
 	file = nil
-}
-func newSheetWithStreamWriter() {
-	file := NewFile()
-	streamWriter, err := file.NewStreamWriter("Sheet1")
-	if err != nil {
-		fmt.Println(err)
-	}
-	for rowID := 1; rowID <= 1000; rowID++ {
-		cell, _ := CoordinatesToCellName(1, rowID)
-		if err := streamWriter.SetRow(cell, []interface{}{rowID}); err != nil {
-			fmt.Println(err)
-		}
-	}
-	if err := streamWriter.Flush(); err != nil {
-		fmt.Println(err)
-	}
-	if err := file.SaveAs("Book1.xlsx"); err != nil {
-		fmt.Println(err)
-	}
 }
 
 func BenchmarkFile_SaveAs(b *testing.B) {