# PR Details

1.align xlsxRow struct.
2.optimization workSheetWriter
3. add NewSheetWithRowNum 

## Description

Mainly to optimize memory allocation

## Related Issue
l wish somebody can optimize xlsxC struct. it's too bigger.Maybe it can be optimized smaller.

## Motivation and Context

sometime l use this package. memory will alloc 500m ~ 1000m. it's too expensive

## How Has This Been Tested

```

go test -run=none -bench=BenchmarkNewSheetWithRowNum -benchmem -count 10

name                   old time/op    new time/op    delta
NewSheetWithRowNum1-8     667µs ± 5%     679µs ±15%     ~     (p=0.968 n=9+10)

name                   old alloc/op   new alloc/op   delta
NewSheetWithRowNum1-8     643kB ± 0%     588kB ± 0%   -8.46%  (p=0.000 n=10+10)

name                   old allocs/op  new allocs/op  delta
NewSheetWithRowNum1-8     10.3k ± 0%      9.2k ± 0%  -10.68%  (p=0.000 n=10+10)


go test -run=none -bench=BenchmarkFile_SaveAs -benchmem -count 10
name           old time/op    new time/op    delta
File_SaveAs-8     738µs ±13%     698µs ±10%     ~     (p=0.105 n=10+10)

name           old alloc/op   new alloc/op   delta
File_SaveAs-8     643kB ± 0%     588kB ± 0%   -8.46%  (p=0.000 n=10+10)

name           old allocs/op  new allocs/op  delta
File_SaveAs-8     10.3k ± 0%      9.2k ± 0%  -10.67%  (p=0.000 n=9+10)



```



## Types of changes

- [ x ] New feature (non-breaking change which adds functionality)
- [ x ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist

- [x ] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ x ] I have read the **CONTRIBUTING** document.
- [ x ] I have added tests to cover my changes.
- [ x ] All new and existing tests passed.