@Theodoree 2020-11-03
go test -run=none -bench=BenchmarkNewSheet -benchmem -count 10

name                        time/op
NewSheetWithRowNum-8         631µs ±12%
NewSheetWithOutRowNum-8      636µs ± 2%
NewSheetWithStreamWriter-8  1.07ms ± 3%

name                        alloc/op
NewSheetWithRowNum-8         508kB ± 0%
NewSheetWithOutRowNum-8      616kB ± 0%
NewSheetWithStreamWriter-8   880kB ± 4%

name                        allocs/op
NewSheetWithRowNum-8         9.18k ± 0%
NewSheetWithOutRowNum-8      9.20k ± 0%
NewSheetWithStreamWriter-8   12.6k ± 0%