@Theodoree 2020-11-03
l will Think about it, maybe I'll delete this api.  `NewSheetWithRowNum`