@xuri 2020-11-03
FYI, [here is](https://xuri.me/excelize/zh-hans/performance.html) a graph that shows the performance comparison of normal API and streaming writing API. I'll public testing code later.