@Theodoree 2020-11-03
@xuri  L have deleted that api. l was testing. l find `NewStreamWriter `  will save more mem.but Execution time will be longer.