@xuri 2020-11-03
Usually, the streaming API will use less memory and faster than normal API, but affected by business scenario, you need trade-off by yourself. Try profiling your program by `pprof` to get more details.